back/requirements.txt
django
psycopg2-binary
django-cors-headers
djangorestframework
pillow

